+++ R/period.R
@@ -14,34 +14,39 @@
 #' @noRd
 parsePeriodDate <- function(x, us.format = NULL)
 {
+    ## Need to be more strict with separators if no days in the periods because false
+    ## positives are more likely e.g. dd-mm-yy could parsed as the period mm-mm-yy
+    ## if '-' is allowed for both separators
     ## \p{Pd} is unicode dash punctuation property
-    date.parts.sep.q <- "[[:space:]]*[,/\\p{Pd}][[:space:]]*"
-    ## allow space as separator for weekly periods since extra date parts make false
-    ## positives very unlikely
-    date.parts.sep.w <- "[[:space:]]*[,/ \\p{Pd}][[:space:]]*"
-    period.sep <- "[[:space:]]*[,/ \\p{Pd}][[:space:]]*"
+    date.parts.sep.no.day <- "[[:space:]]+"
+    period.sep.no.day <- "[[:space:]]*[,/\\p{Pd}][[:space:]]*"
+
+    date.parts.sep.with.day <- "[[:space:]]*[,/ \\p{Pd}][[:space:]]*"
+    period.sep.with.day <- "[[:space:]]*[,/ \\p{Pd}][[:space:]]*"
 
     m.or.b.month <- bOrMMonthRegexPatt()
     year <- yearRegexPatt()
 
-    quarter.regex <- paste0("^", m.or.b.month, date.parts.sep.q,
-                            m.or.b.month, " ", year, "$")
+    opt.year <- paste0("(", date.parts.sep.no.day, year, ")?")
+    quarter.regex <- paste0("^", m.or.b.month, opt.year, period.sep.no.day,
+                            m.or.b.month, date.parts.sep.no.day, year, "$")
 
     ## quarter.regex <- "^[[:alpha:]]{3}-[[:alpha:]]{3} [[:digit:]]{2}$"
     ## # e.g.: 1/02/1999-8/02/1999
     ## week.regex <- paste0("^[[:digit:]]{1,2}/[[:digit:]]{1,2}/[[:digit:]]{4}-",
     ##                      "[[:digit:]]{1,2}/[[:digit:]]{1,2}/[[:digit:]]{4}$")
 
-    dd.mm.yyyy <- dayMonthYearRegexPatt(date.parts.sep.w)
-    mm.dd.yyyy <- monthDayYearRegexPatt(date.parts.sep.w)
-    week.regex.int <- paste0("^", dd.mm.yyyy, period.sep,
+    dd.mm.yyyy <- dayMonthYearRegexPatt(date.parts.sep.with.day)
+    mm.dd.yyyy <- monthDayYearRegexPatt(date.parts.sep.with.day)
+    week.regex.int <- paste0("^", dd.mm.yyyy, period.sep.with.day,
                              dd.mm.yyyy, "$")
-    week.regex.us <- paste0("^", mm.dd.yyyy, period.sep, mm.dd.yyyy, "$")
+    week.regex.us <- paste0("^", mm.dd.yyyy, period.sep.with.day, mm.dd.yyyy, "$")
 
     result <- NA
     ## Q quarters, e.g.: Apr-Jun 08
     if (grepl(quarter.regex, x[1L], perl = TRUE, ignore.case = TRUE))
-        result <- quarterlyPeriodsToDate(x, sep = date.parts.sep.q)
+        result <- quarterlyPeriodsToDate(x, period.sep = period.sep.no.day,
+                                         date.part.sep = date.parts.sep.no.day)
     else
     {
         is.weekly <- FALSE
@@ -52,7 +57,7 @@ parsePeriodDate <- function(x, us.format = NULL)
             is.weekly <- is.weekly || grepl(week.regex.int, x[1L], perl = TRUE,
                                                   ignore.case = TRUE)
         if (is.weekly)
-            result <- weeklyPeriodsToDate(x, us.format, date.parts.sep.w)
+            result <- weeklyPeriodsToDate(x, us.format, date.parts.sep.with.day)
     }
 
     if (any(is.na(result)))
@@ -112,22 +117,31 @@ PeriodNameToDate <- function(x, by, us.format = NULL)
 
 #' Convert Q-quarterly date formats to R date objects
 #'
-#' Parses a character vector of date intervals in Q-quarterly
-#' date format to obtain the start of the interval in R date-time objects
+#' Parses a character vector of date intervals in Q-quarterly date
+#' format to obtain the start of the interval in R date-time objects
 #' @param x Character vector assumed to have elements in date format
-#' "%b-%b %y"; e.g. Apr-Jun 08
-#' @param sep Character string specifying how months in the period are separated
-#' @return Vector containing the start of each period parsed to date objects
+#'     "%b-%b %y"; e.g. Apr-Jun 08.
+#' @param period.sep Character string specifying how the start and end
+#'     points of the period are separated in the elements of \code{x}.
+#' @param date.part.sep Character string specifying how months are
+#'     separated from years in the elements of \code{x}.
+#' @return Vector containing the start of each period parsed to date
+#'     objects
 #' @noRd
 #' @importFrom lubridate dmy year year<-
-quarterlyPeriodsToDate <- function(x, sep = "[/-]")
+quarterlyPeriodsToDate <- function(x, period.sep = "[/-]", date.part.sep = "[[:space:]]")
 {
-    x.split <- strsplit(x, sep, perl = TRUE)
-    start.mon <- vapply(x.split, `[`, 1L, FUN.VALUE = "")
+    x.split <- strsplit(x, period.sep, perl = TRUE)
+    start.dat <- vapply(x.split, `[`, 1L, FUN.VALUE = "")
     end.dat <- vapply(x.split, `[`, 2L, FUN.VALUE = "")
     end.yr <- regmatches(end.dat, regexpr("([0-9]{2}){1,2}$", end.dat))
-    start <- paste0("01-", start.mon, end.yr)
+
+    opt.year <- paste0(date.part.sep, yearRegexPatt())
+    has.start.yr <- grepl(opt.year, start.dat)
+    start <- paste0("01-", start.dat)
+    start[!has.start.yr] <- paste0(start[!has.start.yr], "-", end.yr)
     end <- paste0("01-", end.dat)
+
     start.dmy <- dmy(start, quiet = TRUE)
     end.dmy <- dmy(end, quiet = TRUE)
     ## Need to handle cases where start year < end year