+++ inst/NEWS.md
@@ -4,7 +4,7 @@ Changes in Version 2.4.0 (2017-11-24)
 NEW FEATURES
 
 * `AsDate` gains additional support for date ranges/periods for
-example `"July 7, 1998 - Aug 10, 1999"`
+example `"July 7, 1998 - Aug 10, 1999"` and `"Sep 10 - Jan 11"`
 
 Changes in Version 2.3.2 (2017-11-23)
 --------------------------------------------------------